Optimise and cleanup code (#149)
