+++ src/main/java/uk/gov/hmcts/opal/authentication/util/RequestMatcher.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.common.util;
+package uk.gov.hmcts.opal.authentication.util;
 
 import jakarta.servlet.http.HttpServletRequest;
