+++ src/test/java/uk/gov/hmcts/opal/authentication/config/internal/AuthConfigurationPropertiesStrategyTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.authentication.component.impl;
+package uk.gov.hmcts.opal.authentication.config.internal;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
