+++ src/test/java/uk/gov/hmcts/opal/authentication/service/SessionServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.authentication.service.impl;
+package uk.gov.hmcts.opal.authentication.service;
 
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.mockito.Mockito.verify;
@@ -10,17 +10,16 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import uk.gov.hmcts.opal.authentication.component.SessionCache;
 import uk.gov.hmcts.opal.authentication.model.Session;
 
 @ExtendWith(MockitoExtension.class)
-class SessionServiceImplTest {
+class SessionServiceTest {
 
     @Mock
     private SessionCache sessionCache;
 
     @InjectMocks
-    private SessionServiceImpl sessionService;
+    private SessionService sessionService;
 
     @Test
     void testGetSession() {
